Changed the whole repo
This repository is no longer to host my PHP projects because, upon some digging, I discovered that GitHub only supports static pages and PHP is a non-static language so any website developed for it will inherently be a non-static page. This repository will now be for my future website. I will program that website using Javascript if needed when I need to. In the meantime, it will simply show a hello world script in HTML.